dc_microgrid_dwt/src/agents/processing/noise_filter.py
import numpy as np
import pywt
from src.framework.base_agent import BaseAgent
from src.domain.events import DWTCoefficientsEvent

class NoiseFilterAgent(BaseAgent):
    def setup(self):
        self.subscribe(DWTCoefficientsEvent, self.on_coeffs)

    def on_coeffs(self, event: DWTCoefficientsEvent):
        # In a full implementation, this might modify coefficients and re-publish
        # Or it might just log noise levels.
        # For this high-speed architecture, we might just monitor noise floor.
        pass
